No, Larry apparently really intended assignment here.
Use of the LGTM code analyzer
(https://lgtm.com/projects/g/Perl/perl5/alerts/?mode=tree&ruleFocus=2158670641)
flagged two places in regcomp.c for "Use of '=' where '==' may have been
intended."  However, when we change '=' to '==' in two places -- one of
which has been in Perl 5 since commit 0 in 1994 -- we end up with
massive failures in t/re/fold_grind.t.

So this commit is for demonstration purposes only and should not be
merged into blead.